RPM: move supports_arch() method to Spec class
Do not abstract RPM spec file load and spec attributes check behind
Package class. Parsing spec file can fail, and error raised due to spec
file parsing failure in Package.supports_arch() is arguably an API
smell. It feels better the caller initializes its own instance of the
Spec class and runs supports_arch() method on this instance in order to
decorrelate parsing and attributes checks.

Tests are updated accordingly.
